+++ python_modules/dagster/dagster/core/types/__init__.py
@@ -1,7 +1,7 @@
 from .builtin_enum import BuiltinEnum
 from .wrapping import Nullable, List
-from .config import Dict, NamedDict, Selector
 from .field import Field
+from .field_utils import Dict, NamedDict, Selector, NamedSelector
 from .runtime import PythonObjectType
 
 Any = BuiltinEnum.ANY
